pyproject.toml
[project]
name = "microservice-qa"
version = "0.1.0"
description = ""
authors = [
    { name = "Dimitri Maksimov", email = "salesakk@gmail.com" }
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "fastapi[standard] (>=0.120.2,<0.121.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "fastapi-pagination (>=0.15.0,<0.16.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "jsondiff (>=2.2.1,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "faker (>=37.12.0,<38.0.0)",
    "curlify (>=3.0.0,<4.0.0)"
]

[tool.pytest.ini_options]
python_files = ["*_tests.py", "test_*.py"]
addopts = ["-v", "-s"]
pythonpath = '.'


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
